♻️ Refacotr: seperate image utils
이미지 관련 함수 분리
